feat#1896: move test-gen exception into dedicated package to appease arch rules
