+++ lib/kadai-test-api/src/main/java/io/kadai/testapi/generator/exceptions/TestDataGenerationException.java
@@ -16,7 +16,7 @@
  *
  */
 
-package io.kadai.testapi.generator;
+package io.kadai.testapi.generator.exceptions;
 
 /** Runtime exception for generation or persistence failures. */
 public class TestDataGenerationException extends RuntimeException {
